Implement a wireless streaming service on ugglan.
The c++ server is fully implemented and a basic python client
exists. Also made a lot of refactoring. A selection of it follows.

- Add header guards to gzip lib and throw error if empty input
- Rename c++ utils to common_utils to avoid name clashing
- Refactor data log e.g., move logic to data_log_utils
- Make use of size_t where applicable
- Throw c++ style runtime error in base64 lib, rm new line by default

+++ raspi/inc/serial_conn.h
@@ -27,8 +27,8 @@ public:
     virtual bool open(Mode mode, ControlFlags flags);
     virtual bool close();
 
-    virtual uint32_t bytes_available();
-    virtual uint32_t read(uint8_t* buf, uint32_t size);
+    virtual size_t bytes_available();
+    virtual size_t read(uint8_t* buf, size_t size);
 private:
     int8_t _fd;
     std::string _device;
