+++ raspi/tests/serial_conn_stub.cpp
@@ -1,6 +1,6 @@
 #include <serial_conn_stub.h>
 
-void SerialConnStub::set_read_buf(uint8_t* buf, uint32_t size)
+void SerialConnStub::set_read_buf(uint8_t* buf, size_t size)
 {
     _read_buf_size = size;
 
@@ -18,14 +18,14 @@ bool SerialConnStub::close()
     return true;
 }
 
-uint32_t SerialConnStub::bytes_available()
+size_t SerialConnStub::bytes_available()
 {
     return _read_buf_size;
 }
 
-uint32_t SerialConnStub::read(uint8_t* buf, uint32_t size)
+size_t SerialConnStub::read(uint8_t* buf, size_t size)
 {
-    uint32_t bytes_to_read = std::min(_read_buf_size, size);
+    size_t bytes_to_read = std::min(_read_buf_size, size);
     std::memcpy(buf, _read_buf, bytes_to_read);
 
     return bytes_to_read;
